@@ -1,14 +1,11 @@
-﻿using System.Collections.Generic;
+﻿namespace LeaderAnalytics.LeaderPivot;
 
-namespace LeaderAnalytics.LeaderPivot
+/// <summary>
+/// A row on a matrix.  A row has a theoretically unlimited number of columns.
+/// </summary>
+public class MatrixRow
 {
-    /// <summary>
-    /// A row on a matrix.  A row has a theoretically unlimited number of columns.
-    /// </summary>
-    public class MatrixRow
-    {
-        public List<MatrixCell> Cells { get; set; }
+    public List<MatrixCell> Cells { get; set; }
 
-        public MatrixRow() => Cells = new List<MatrixCell>();
-    }
+    public MatrixRow() => Cells = new List<MatrixCell>();
 }
